move pagetable from buffermgr.go
